// packages/interpreter/src/token-type.ts
export enum TokenType {
  NOT = "NOT",
  NOT_EQUAL = "NOT_EQUAL",
  EQUAL = "EQUAL",
  EQUAL_EQUAL = "EQUAL_EQUAL",
  GREATER_THAN = "GREATER_THAN",
  GREATER_THAN_EQUAL = "GREATER_THAN_EQUAL",
  LESS_THAN = "LESS_THAN",
  LESS_THAN_EQUAL = "LESS_THAN_EQUAL",
  PLUS = "PLUS",
  MINUS = "MINUS",
  STAR = "STAR",
  SLASH = "SLASH",
  AND = "AND",
  OR = "OR",
  IF = "IF",
  ELSE = "ELSE",
  THEN = "THEN",
  END = "END",
  IS = "IS",
  EQUALS = "EQUALS",
  DEFINE = "DEFINE",
  REPEAT = "REPEAT",
  TIMES = "TIMES",
  EACH = "EACH",
  IN = "IN",
  WITH = "WITH",
  COMMA = "COMMA",

  // literals
  STRING_LITERAL = "STRING_LITERAL",
  NUMBER_LITERAL = "NUMBER_LITERAL",
  TRUE_LITERAL = "TRUE_LITERAL",
  FALSE_LITERAL = "FALSE_LITERAL",
  VARIABLE_IDENTIFIER = "VARIABLE_IDENTIFIER",
  SELECTOR_IDENTIFIER = "SELECTOR_IDENTIFIER",
  IDENTIFIER = "IDENTIFIER",

  // commands
  GO = "GO",
  TO = "TO",
  BACK = "BACK",
  FORWARD = "FORWARD",
  REFRESH = "REFRESH",
  BE = "BE",
  VISIBLE = "VISIBLE",
  HIDDEN = "HIDDEN",
  CONTAIN = "CONTAIN",
  ENABLED = "ENABLED",
  DISABLED = "DISABLED",
  CHECKED = "CHECKED",
  CHECK = "CHECK",
  UNCHECK = "UNCHECK",
  PAGE = "PAGE",
  TITLE = "TITLE",
  URL = "URL",
  THE = "THE",
  ALL = "ALL",
  LAST = "LAST",
  FROM = "FROM",
  TYPE = "TYPE",
  CLICK = "CLICK",
  PRESS = "PRESS",
  WAIT = "WAIT",
  SELECT = "SELECT",
  OPEN = "OPEN",
  CLOSE = "CLOSE",
  HOVER = "HOVER",
  SCROLL = "SCROLL",
  ASSERT = "ASSERT",
  EXPECT = "EXPECT",
  CALL = "CALL",
  SET = "SET",
  INTO = "INTO",
  PRINT = "PRINT",
  AS = "AS",
  GET = "GET",
  HAVE = "HAVE",
  TEXT = "TEXT",
  VALUE = "VALUE",
  FOR = "FOR",
  SECOND = "SECOND",
  SECONDS = "SECONDS",
  LOAD = "LOAD",
  APPEAR = "APPEAR",
  DISAPPEAR = "DISAPPEAR",
  ORDINAL = "ORDINAL",
  OVER = "OVER",

  EOF = "EOF",
}